refactor(dependencies): move DependencyManager into a submodule :recycle:
